implemented algorithms
text normalization algorithm with TextNormalization, WhitespaceNormalization, RepetitionNormalization, InvisibleCharacterNormalization functions. Implemented leetspeak normalization with LeetspeakNormalization and IsPrimarlyCyrillic functions. Implemented word checker with Tokenization and Hash-Trie building with further checking with Wordchecking and Tokenize functions. Currently they are NOT adapted to the moderation-service environment and require further adjustments and combination with proto and with the database